somme support for utf-8 in latex
This allow the use of some utf-8/unicode in the ipynb->latex conversion.

tested in codecell comment and markdown.

Most of the caracter with accent, and thins like squareroot symbol and
so on find their way through until the .tex file, but pdlatex
compilation only works out of the box only with accentuated letters