added 'py' format, which is cleaner than PyWriter
additionally, it allows for future extensions (where you drop the prompt
numbers, or all of the markdown cells, or drop just the code and keep
only the markdown around as comments, as a template for homework, for
example).

Currently, as noted in ConverterPy's docstring, it does not guarantee
that what gets written is valid python, since the code cells may contain
magics.