Fixing issue preventing the correct read of images by full_html and reveal exporters.
Currently to see and images in the notebook inside a md cell you can use the prefix files/ to read the content of the directory where the ipynb lives. But, if you do a conversion using full_html or reveal exporters, the are not rendered ok because the previous version of this function erased `/files/` instead of `files/`. With this fix, the html files produced by full_html and reveal exporters can read the images in a correct way.