Rehash and simplify slide export
This is a medium size reworking of the slide export code. The major
changes are:

1. Got rid of the RevealHelpPreprocessor, merging its functionality into
SlideExporter. Neither is useful without the other, so it doesn't make
sense to have a preprocessor that can be turned on and off separately.
This slightly breaks config: RevealHelpPreprocessor.url_prefix becomes
SlidesExporter.reveal_url_prefix. I can do a compatibility shim if
needed, but I suspect it's not worth the added complexity.

2. The template and the cell metadata added to slides to facilitate
grouping them into slides/subslides/fragments are (I hope) simplified.
It now has six keys: `{slide,subslide,fragment}_{start,end}` (plus
slide_type). In the template, we just prefix & suffix the cell with
appropriate opening and closing tags according to the start/end
attributes.

3. The extra metadata is now added to a copy of the notebook, rather
than modifying the notebook object. This has no effect when using
nbconvert at the command line, but if you use it in a Python script,
this means the notebook object you pass in is not changed.