Make 'element' available for JS output to add to document
From discussion with Min. JS output in the notebook can add to 'element'
to put things on the page in that output area. This adds the same
variable for use by JS output in nbconverted HTML.

I'm playing with this in altwidgets-prototype, and it seems to work. I
haven't thought through all the possible scoping issues, though.