Update docs readme
Pursuant to https://github.com/jupyter/nbconvert/issues/379, this seemed like a good thing to make as a quick fix, with the more substantial restructuring to occur later. 

Notably this requires using conda, which not all users may rely on. I do not know if there is a way to expose the `environment.yml` file to a `pip` command for non-conda users, but it seems like it shouldn't be a long run requirement that you use `conda` in order to build our docs.