First pass at an expected failure test
add import

add class definition

relocate to surface at top level

Test result

clean up code

Use other configuration method

try a plain raise

start writing direct debugging code, structure seems to be in place

This should actually fail

Nope, everything got worse

Here's the error report, note the object definition is not ExtensionTolerantLoader

Raise the right exceptiontype

Change it around to directly set the loader

test result failure

This is failing where I wouldn't expect it to fail

Ok, now we're getting somewhere. Now it seems to be an issue with the conditional which was never being hit.

Approaches traitlets incorrectly, but now we may in the ballpark

Now it's happening in the Choice Loader and not the ExtensionTolerantLoader

check that the conditional is passing. satifice for tonight.