Allow kernel authors to suggest a package for script export
Part of gh-361

Now that we use entry points to find exporters, kernel authors could
provide a custom script exporter, and specify it in language_info
metadata. But currently, if that package is not installed, nbconvert
will fail with an unhelpful error message. This is a problem especiall
for kernels which are not distributed as Python packages (e.g. R, Julia)

With this PR, if the specified exporter cannot be found, nbconvert will
display a warning and fall back on plain script export. It also looks
for a new field in language_info, called 'nbconvert_exporter_dist', to
provide a suggestion of a package to install to provide better export.