Add width tags for jpeg and png when present
The display(Image(..., retina=2)) command creates metadata about the image explaining the size that it should be displayed at in the notebook.  This information, if present, can be carried into the notebook and converted to rst for use in documenters such as Sphinx.  If not present, retina images appear in .rst (and from there to other formats such as HTML) as twice their intended size.