Move preprocess logic into setup_preprocessor and start_new_kernel
setup_preprocessor is a contextmanager that handles setting and cleaning
up attributes needed by the rest of the execution logic (specifically
self.nb, self.km, self.kc).

start_new_kernel was previously a callback function and is now a
separate method used inside setup_preprocessor.

This also introduces a dynamic default for the `kernel_name` traitlet
that checks if self.nb is defined to see if it has a kernel name, and
otherwise will raise an AttributeError (since that means it is being
relied upon without any way to know which kernel is to be used.