add new find_kernel_name API for finding kernel names from nb metadata
This is being introduced because we previously allowed users to pass ""
to indicate that the kernel name should be exclusively inferred from the
notebook itself. Doing so now overrides the default behaviour (which now
defaults to infering the kernel name from the notebook itself if nothing
is set). This is not a great pattern since it involves setting a
parameter to a magic value to keep the default behaviour.

This also issues a warning if you use the old way to do this.