Change RegExRemove to remove code cells with output
Previously, the RegExRemovePreprocessor treated code cells without outputs specially: they were never removed, even if they matched the RegEx. This behavior was intentional but only documented in an internal method, check_conditions, not in the docs of an external method. See Issue #1091 for discussion.

This commit removes that behavior and updates the documentation of check_conditions. It further removes the test that checked for this behavior.