Fix Inkscape executable for macOS & Inkscape v1.0
Starting with Inkscape v1.0 (currently beta), the executable for Inkscape (`inkscape`) can now be found inside the `Inkscape.app/Contents/MacOS` directory. Note that this file might also be present on versions prior to 1.0.
Thus, the order in which we check the presence of the executable is important. If the old `INKSCAPE_APP` is present, it should be preferred over the one in `Inkscape.app/Contents/MacOS`.

I haven't got the time and resources to test this on Inkscape v0.92.4 as there is
 > no .dmg file for macOS for the current Inkscape version at this time
and Inkscape v0.92.2 is unsupported on macOS 10.15. I was only able to test this fix with the current beta `v1.0beta2` of Inkscape on macOS.