Move disable logic to get_exporter
This commit moves the logic for disabling exporters into the
get_exporter function rather than just updating the list of available
exporters. If someone explicitly asks for an exporter which is disabled
in their configuration they will get a ExporterDisabledError exception.
The logic in get_exporter_names can be simplified so it only has to
catch (and ignore) that exception when building the list of enabled
exporters.

Signed-off-by: Ian Allison <iana@pims.math.ca>