DOC: Autoreformat all the docstring to match the numpydoc format.
This uses a prototype project of mine, so might be incomplete, and try
to be better at understanding (and reformatting, docstrings complient
with numpydoc.

I understand that you would be opposed to large scale reformatting.

This is on top of #1323 which made numpydoc choke on the docstring and
prevented correct reformatting.