Allow passing extra args to code highlighter
I primarily want to allow for the possibility of line numbers
in HTML output. Pygments supports this natively, but we don't
have a way to pass arguments through to pygments.

A traitlet is added to allow passing arbitrary extra args
to the Pygments formatter, for Html and LaTeX. Should allow for
line number control and much more :)

This has been something folks have asked for for a while.

Ref https://github.com/ipython/ipython/issues/5847
Ref https://github.com/jupyter/nbconvert/issues/427
Ref https://github.com/notebook-sharing-space/nbss/issues/34