Set cm-unicode font for latex if available
For LuaLaTeX and XeLaTeX, the default fonts usually don't support
Greek text, causing problems when Greek letters are used outside of
math mode (eg, as text or as symbols in code).  This change to the
base template uses the cm-unicode fonts, if they are available.
Fonts are left unchanged if cm-unicode is not available, in which
case Greek letters in math will still work, but text will be missing.

Loading is done here using cm-unicode's filenames, rather than the
fonts' symbolic names, because XeLaTeX cannot search for symbolic
names of fonts that are not available to the system (but may be
available to LaTeX).  LuaLaTeX does not have this limitation, but can
also load using filenames.