Update nbconvert_library.ipynb
The generated html file was not following the 'classic' template, but instead 'lab', when I ran the code according to documentaion. But when I passed the template_name as a keyword arguement to HTMLExporter it worked. So I made necessary changes to the code, in a way which worked for me. Please do a review of the code and accept the changes